Implement backtracking using a iterator
The "match constructor" operation was returning a single matching
constructor, it was modified to return a iterator over all the
matching constructors.

This way is possible to backtrack the constructor matching without
needing to save/restoring the current constructor offset, this logic
could be handled by the `Matcher`.

Other changes was required to satisfy the borrow checker:

* Created `InnerDecoder` to handle "mut" and "non-mut" references,
* The "get_token" logic was moved to `Token`,
* Match constructors now receive raw data instead of a ref to `Decoder`.

The only change to the API was made to `Decoder::global_context` that
can't be read/write directly, now it can only be accessed by
`Decoder::context` and `Decoder::set_context`.